Larger counts

+++ benches/benches.rs
@@ -11,17 +11,17 @@ fn complex_function(list: &[f64; 5], _: Option<Arc<()>>) -> f64 {
 }
 
 // Timeout iterations
-const LIMIT: u32 = 1000000; // 1 million
-const GRID_SIMPLE_LIMIT: u32 = 100; // LIMIT.powf(1. / 3.);
-const GRID_COMPLEX_LIMIT: u32 = 15; // LIMIT.powf(1. / 5.);
-const SIMULATED_ANNEALING_LIMIT: (f64, f64, u32) = (100., 1., 1000); // = 100 * 10000
+const LIMIT: u64 = 1000000; // 1 million
+const GRID_SIMPLE_LIMIT: u64 = 100; // LIMIT.powf(1. / 3.);
+const GRID_COMPLEX_LIMIT: u64 = 15; // LIMIT.powf(1. / 5.);
+const SIMULATED_ANNEALING_LIMIT: (f64, f64, u64) = (100., 1., 1000); // = 100 * 10000
 
 const SIMPLE_EXIT: f64 = 18.;
 const COMPLEX_EXIT: f64 = -17.;
 
-const BIG_LIMIT: u32 = 10000000; // 10 million
-const GRID_BIG_LIMIT: u32 = 215; // BIG_LIMIT.powf(1. / 3.);
-const SIMULATED_ANNEALING_BIG_LIMIT: u32 = BIG_LIMIT / 100; // BIG_LIMIT / 100;
+const BIG_LIMIT: u64 = 10000000; // 10 million
+const GRID_BIG_LIMIT: u64 = 215; // BIG_LIMIT.powf(1. / 3.);
+const SIMULATED_ANNEALING_BIG_LIMIT: u64 = BIG_LIMIT / 100; // BIG_LIMIT / 100;
 
 use simple_optimization::Polling;
